+++ src/components/blog/BlogAuthor.js
@@ -47,7 +47,7 @@ const BlogAuthor = () => {
       <div className="portrait">
         <Img
           fluid={myData.portrait.childImageSharp.fluid}
-          title="Jacob D. Castro"
+          title="Boo & Bae"
           alt="photo from side of me looking left with sunset beach in background"
           style={{
             height: '80px',