+++ src/components/layout/Header.js
@@ -28,7 +28,7 @@ const Header = () => {
         <div id="logo">
           <Link to="/" aria-label="to home page">
             <LogoSVG />
-            <h2>JACOB D. CASTRO</h2>
+            <h2>Boo & Bae</h2>
           </Link>
         </div>