+++ src/pages/contact.js
@@ -12,7 +12,7 @@ const Contact = ({ path, data }) => {
     description:
       // eslint-disable-next-line quotes
       "Let's chat! I'm open to lots of new opportunities, freelance or a fulltime position.",
-    url: 'https://jacobdcastro.com/contact',
+    url: 'https://boobae.blog/contact',
     imgUrl: `${data.pageImg.publicURL}`,
     imgAlt:
       'jdcastro logo, twitter, instagram, facebook, github icons with @jacobdcastro username',